frontend: feature based tests
