+++ frontend/playwright/home.spec.ts
@@ -1,0 +1,12 @@
+import { expect, test } from "@playwright/test"
+import { HomePage } from "./pages/HomePage.ts"
+
+test("shows the public landing page", async ({ page }) => {
+  const homePage = await HomePage.goto(page)
+
+  await test.step("Verify the landing page content", async () => {
+    await expect(homePage.heading).toBeVisible()
+    await expect(homePage.playForFreeLink).toBeVisible()
+    await expect(homePage.createAccountLink).toBeVisible()
+  })
+})
